refactor: renderer (#3)
* refactor: reorg roma engine

* refactor: minor improvments

* chore: refinement

* fix: z-sorting

* fix: world y

+++ crates/definitions/src/types/image.rs
@@ -1,9 +1,9 @@
 #[derive(Debug, Clone, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
 pub struct Image {
     pub id: usize,
-    pub file_num: u32,
-    pub x: u16,
-    pub y: u16,
-    pub width: u16,
-    pub height: u16,
+    pub file_num: usize,
+    pub x: usize,
+    pub y: usize,
+    pub width: usize,
+    pub height: usize,
 }
